requirements.txt
Django
djangorestframework
psycopg2-binary
python-dotenv
requests
yadisk
pysmb
python-dateutil
pysrt
croniter
celery[redis]
ffmpeg-python
django-celery-beat
django-celery-results
gunicorn
faster-whisper
django-filer
easy-thumbnails